refactor:onboarding配下の型定義を集約
