+++ src/features/onboarding/components/onboarding-character.tsx
@@ -1,4 +1,5 @@
 import { Button } from "@/components/ui/button";
+import type { OnboardingCharacterProps } from "@/features/onboarding/types/types";
 import { motion } from "framer-motion";
 import { ChevronDown } from "lucide-react";
 import Image from "next/image";
@@ -9,13 +10,6 @@ import {
 } from "../constants/constants";
 import { onboardingDialogues } from "../constants/onboarding-texts";
 import { getButtonText, isFinalScreen, sanitizeHtml } from "../utils/utils";
-
-interface OnboardingCharacterProps {
-  currentDialogue: number;
-  isAnimating: boolean;
-  onNext: () => void;
-  onScrollDown: () => void;
-}
 
 /**
  * オンボーディングキャラクターコンポーネント（期日前投票専用）
