+++ src/features/onboarding/hooks/use-onboarding-state.ts
@@ -1,11 +1,11 @@
+import type {
+  UseOnboardingActions,
+  UseOnboardingState,
+} from "@/features/onboarding/types/types";
 import { useEffect, useRef, useState } from "react";
 import { ANIMATION_DURATION } from "../constants/constants";
 import { onboardingDialogues } from "../constants/onboarding-texts";
-import type { UseOnboardingActions, UseOnboardingState } from "../types/types";
-import {
-  calculateDefaultScrollPosition,
-  calculateScrollPosition,
-} from "../utils/utils";
+import { calculateDefaultScrollPosition } from "../utils/utils";
 
 /**
  * オンボーディング状態管理フック（期日前投票専用）
