+++ src/features/onboarding/types/types.ts
@@ -1,3 +1,6 @@
+import type { Tables } from "@/lib/types/supabase";
+import type { ReactNode } from "react";
+
 /**
  * オンボーディング機能で使用する型定義（期日前投票専用）
  */
@@ -26,6 +29,37 @@ export interface OnboardingModalProps {
   onOpenChange: (open: boolean) => void;
 }
 
+// オンボーディングボタンのProps
+export interface OnboardingButtonProps {
+  children?: ReactNode;
+  variant?: "default" | "outline" | "link";
+  className?: string;
+}
+
+// オンボーディングキャラクターのProps
+export interface OnboardingCharacterProps {
+  currentDialogue: number;
+  isAnimating: boolean;
+  onNext: () => void;
+  onScrollDown: () => void;
+}
+
+// オンボーディングミッションカードのProps
+export interface OnboardingMissionCardProps {
+  mission: Omit<Tables<"missions">, "slug">;
+  achieved: boolean;
+  achievementsCount?: number;
+  userAchievementCount?: number;
+  onCardClick?: () => void;
+}
+
+// オンボーディングミッション詳細のProps
+export interface OnboardingMissionDetailsProps {
+  mission: MockMission;
+  isSubmissionCompleted: boolean;
+  onSubmit: () => void;
+}
+
 // フック型
 export interface UseOnboardingState {
   currentDialogue: number;
@@ -39,4 +73,13 @@ export interface UseOnboardingActions {
   handleSubmit: () => void;
   handleScrollDown: () => void;
   handleOpenChange: (open: boolean) => void;
+}
+
+// オンボーディングダイアログの型
+export interface OnboardingDialogue {
+  id: number;
+  text: string;
+  isWelcome: boolean;
+  showMissionCard?: boolean;
+  showMissionDetails?: boolean;
 }
